refactor(credentials)!: replace the config loader global with a shared package
config and credentials both need to read the Docker config file, but
credentials could not import config without a cycle. That was worked around
with a process-wide registration hook: config called SetDefaultConfigLoader in
an init, credentials read the global, and anyone who used credentials without
also importing config got ErrNoConfigLoader. Import order decided whether the
library worked, which is why credentials/main_test.go carried a TestMain whose
only job was to diagnose the missing registration.

Move the file format into registry/remote/internal/configfile, holding Config,
Load and the ConfigFile interface. Both packages import it and neither imports
the other for this purpose, so the dependency is expressed in the import graph
rather than in init order.

config.Config stays a type alias, so the exported surface of config is
unchanged for anyone using it to read a config file.

BREAKING CHANGE: SetDefaultConfigLoader, ConfigFileLoader and ErrNoConfigLoader
are removed. Nothing needs to replace them at a call site — credentials now
reads config files directly, so the registration they existed to perform is no
longer necessary.

The TestMain guard in credentials/main_test.go goes with them, and the config
file tests and testdata move alongside the code they exercise.
