+++ registry/remote/internal/configfile/testdata/bad_config
