+++ registry/remote/internal/configfile/testdata/credHelpers_config.json
