+++ registry/remote/internal/configfile/testdata/credsStore_config.json
