+++ registry/remote/internal/configfile/testdata/hierarchical_auths_config.json
