+++ registry/remote/internal/configfile/testdata/invalid_auths_config.json
