+++ registry/remote/internal/configfile/testdata/invalid_auths_entry_config.json
