+++ registry/remote/internal/configfile/testdata/legacy_auths_config.json
