+++ registry/remote/internal/configfile/testdata/no_auths_config.json
