+++ registry/remote/internal/configfile/testdata/valid_auths_config.json
