+++ registry/remote/internal/configfile/testdata/whitespace.json
